+++ src/synth_setter/models/vst_ff_module.py
@@ -85,7 +85,7 @@ class VSTFeedForwardModule(LightningModule):
 
     def setup(self, stage: str) -> None:
         if self.hparams.compile and stage == "fit":
-            self.net = torch.compile(self.net)
+            self.net.compile()
 
     def on_before_optimizer_step(self, optimizer) -> None:
         norms = grad_norm(self.net, 2.0)